fix(tenderize): rename file + updated comment
